moved linq relevant files to proper namespace
